requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
deprecated==1.2.18
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
limits==5.5.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
slowapi==0.1.9
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.47.3
tomli==2.2.1
typing-extensions==4.15.0
typing-inspection==0.4.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
